backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.10
yfinance==0.2.40
pycryptodomex==3.20.0
cryptography==42.0.8
kiteconnect==5.0.1
SharekhanApi==1.0.0
